Split elements into separate packages
Split model and viewcontroller into separate packages. (EG)
