+++ src/main/java/com/cbm/tda367/model/BookDatabase.java
@@ -1,9 +1,6 @@
-package com.cbm.tda367;
+package com.cbm.tda367.model;
 
 
-import javafx.scene.image.Image;
-
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
