+++ src/main/java/com/cbm/tda367/model/ListingDatabase.java
@@ -1,7 +1,6 @@
-package com.cbm.tda367;
+package com.cbm.tda367.model;
 
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 /**
