+++ src/main/java/com/cbm/tda367/viewcontroller/BoughtBooksMiniatureViewController.java
@@ -1,5 +1,7 @@
-package com.cbm.tda367;
+package com.cbm.tda367.viewcontroller;
 
+import com.cbm.tda367.model.ApplicationModel;
+import com.cbm.tda367.model.Listing;
 import javafx.event.Event;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -31,7 +33,7 @@ public class BoughtBooksMiniatureViewController {
         this.listing = listing;
         this.manager = manager;
 
-        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("BoughtBooks.fxml"));
+        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/com/cbm/tda367/BoughtBooks.fxml"));
         fxmlLoader.setRoot(this);
         fxmlLoader.setController(this);
