+++ src/main/java/com/cbm/tda367/viewcontroller/ListingViewController.java
@@ -1,5 +1,7 @@
-package com.cbm.tda367;
+package com.cbm.tda367.viewcontroller;
 
+import com.cbm.tda367.model.ApplicationModel;
+import com.cbm.tda367.model.Listing;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.image.Image;
@@ -37,7 +39,7 @@ public class ListingViewController extends AnchorPane {
         this.listing = listing;
         this.manager = manager;
 
-        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("shop-page-subscription(1).fxml"));
+        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/com/cbm/tda367/shop-page-subscription(1).fxml"));
         fxmlLoader.setRoot(this);
         fxmlLoader.setController(this);
