+++ src/main/java/com/cbm/tda367/viewcontroller/ReservedBooksMiniatureViewController.java
@@ -1,5 +1,7 @@
-package com.cbm.tda367;
+package com.cbm.tda367.viewcontroller;
 
+import com.cbm.tda367.model.ApplicationModel;
+import com.cbm.tda367.model.Listing;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.shape.Rectangle;
@@ -44,7 +46,7 @@ public class ReservedBooksMiniatureViewController {
         this.manager = manager;
         this.listing = listing;
 
-        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("ReservedBooks.fxml"));
+        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/com/cbm/tda367/ReservedBooks.fxml"));
         fxmlLoader.setRoot(this);
         fxmlLoader.setController(this);
