+++ src/main/java/com/cbm/tda367/viewcontroller/SubscribedBooksMiniatureViewController.java
@@ -1,5 +1,7 @@
-package com.cbm.tda367;
+package com.cbm.tda367.viewcontroller;
 
+import com.cbm.tda367.model.ApplicationModel;
+import com.cbm.tda367.model.Book;
 import javafx.event.Event;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -44,7 +46,7 @@ public class SubscribedBooksMiniatureViewController extends AnchorPane {
     public SubscribedBooksMiniatureViewController( Book book,ControllerManager manager) {
         this.book = book;
         this.manager = manager;
-        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("SubscribedBooks.fxml"));
+        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/com/cbm/tda367/SubscribedBooks.fxml"));
         fxmlLoader.setRoot(this);
         fxmlLoader.setController(this);
