+++ src/main/java/module-info.java
@@ -5,4 +5,8 @@ module com.cbm.tda367 {
 
     opens com.cbm.tda367 to javafx.fxml;
     exports com.cbm.tda367;
+    exports com.cbm.tda367.model;
+    opens com.cbm.tda367.model to javafx.fxml;
+    exports com.cbm.tda367.viewcontroller;
+    opens com.cbm.tda367.viewcontroller to javafx.fxml;
 }
